+++ lib/vin_decoder.dart
@@ -2,5 +2,6 @@
 
 library vin_decoder;
 
+export 'src/nhtsa_vehicle.dart';
 export 'src/vin_decoder_base.dart';
 export 'src/vin_generator.dart';